+++ _example/echo/install.sh
@@ -1,3 +1,2 @@
-#!/bin/bash
-
+#!/bin/sh
 go get github.com/labstack/echo/v4