+++ _example/fasthttp/install.sh
@@ -1,6 +1,5 @@
-#!/bin/bash
-
-go get github.com/fasthttp/router@v1.3.12
-go get github.com/soheilhy/cmux@v0.1.5
-go get github.com/valyala/fasthttp@v1.24.0
-go get github.com/valyala/fasthttp/fasthttpadaptor@v1.24.0
+#!/bin/sh
+go get github.com/fasthttp/router
+go get github.com/soheilhy/cmux
+go get github.com/valyala/fasthttp
+go get github.com/valyala/fasthttp/fasthttpadaptor