+++ _example/iris/install.sh
@@ -1,3 +1,2 @@
-#!/bin/bash
-
+#!/bin/sh
 go get github.com/kataras/iris/v12@v12.1.8